+++ Runtime/DebugGroups.cs
@@ -1,0 +1,14 @@
+namespace Pixygon.DebugTool {
+    public enum DebugGroup {
+        Addressable,
+        Tablet,
+        Application,
+        Avatar,
+        Actor,
+        Nft,
+        Property,
+        Camera,
+        PixygonApi,
+        PixygonMicro
+    }
+}