+++ CHANGELOG.md
@@ -4,3 +4,8 @@ Pixygon - Debug Tool
 - Added editor-tools for setting up debugprofile if none is present
 - Added automatic, temporary, creation of debugprofiles if none is present
 - Changed name of package
+
+0.5.2
+- Improved how debug-profiles are created and saved to disk
+- Changed prefix-formatting to be more clear
+- Added tests